requirements.txt
aio-pika==9.2.2
aiormq==6.7.7
asgiref==3.8.1
asyncache==0.3.1
atomicwrites==1.4.1
cachetools==5.3.1
Django==5.1.1
django-admin-sortable2==2.2.3
django-tinymce==4.1.0
furl==2.1.3
idna==3.7
multidict==6.0.5
orderedmultidict==1.0.1
pamqp==3.2.1
pillow==10.4.0
pydantic_core==2.6.1
pyproject_hooks==1.0.0
pytz==2024.1
PyYAML==6.0.1
six==1.16.0
sqlparse==0.5.1
types-cachetools==5.3.0.6
typing_extensions==4.7.1
yarl==1.9.4
